refactor to controller approach, admin navigation, adjust seeds
