+++ helpers/passport.js
@@ -1,3 +1,6 @@
+var passport = require('passport');
+var User = require('../models/user');
+var LocalStrategy = require('passport-local').Strategy;
 module.exports.ensureAuthenticated = (req, res, next) => {
   if (req.isAuthenticated()) {
     return next();
@@ -5,3 +8,38 @@ module.exports.ensureAuthenticated = (req, res, next) => {
     res.redirect("/users/login");
   }
 };
+
+passport.use(
+  new LocalStrategy((username, password, done) => {
+    console.log("GRRR", username)
+    User.getUserByUsername(username, (err, user) => {
+      if (err) throw err;
+      if (!user) {
+        return done(null, false, {
+          message: 'Unknown User'
+        });
+      }
+      User.comparePassword(password, user.password, (err, isMatch) => {
+        if (err) throw err;
+        if (isMatch) {
+          return done(null, user);
+        } else {
+          return done(null, false, {
+            message: 'Invalid password'
+          });
+        }
+      });
+    });
+  })
+);
+
+passport.serializeUser((user, done) => {
+  done(null, user.id);
+});
+
+passport.deserializeUser((id, done) => {
+  User.getUserById(id, (err, user) => {
+    done(err, user);
+  });
+});
+
