+++ routes/admin/categories.js
@@ -1,76 +1,17 @@
-const Category = require("../../models/category");
+const AdminCategoriesController = require("../../controllers/admin/AdminCategoriesController");
 
 const express = require("express");
 const router = express.Router();
 const { ensureAuthenticated } = require('../../helpers/passport')
 
-router.get("/", ensureAuthenticated, async function (req, res) {
-  //here we get the whole collection and sort by order
-  let categories = await Category.find({}).exec();
-  res.render("admin/categories", {
-    categories: categories,
-    message: res.locals.message,
-    color: res.locals.color
-  });
-});
+router.get("/", ensureAuthenticated, AdminCategoriesController.getCategories);
 
-router.get("/:id", ensureAuthenticated, function (req, res) {
-  Category.findById(req.params.id, function (err, category) {
-    res.render("category", {
-      category: category
-    });
-  });
-});
+router.get("/edit/:id", ensureAuthenticated, AdminCategoriesController.editCategory);
 
-router.get("/edit/:id", ensureAuthenticated, function (req, res) {
-  Category.findById(req.params.id, function (err, category) {
-    res.render("editCategory", {
-      category: category,
-      message: res.locals.message,
-      color: res.locals.color
-    });
-  });
-});
+router.post("/", ensureAuthenticated, AdminCategoriesController.createCategory);
 
-router.post("/", ensureAuthenticated, function (req, res) {
-  var category = new Category(); // create a new instance of the category model
-  category.name = req.body.name; // set the categories name (comes from the request)
+router.delete("/delete/:id", ensureAuthenticated, AdminCategoriesController.deleteCategory);
 
-  // save the category and check for errors
-  category.save(function (err) {
-    if (err) res.send(err);
-    console.log("Category created:", category);
-    res.redirect("/categories?alert=created");
-  });
-});
+router.put("/update/:id", ensureAuthenticated, AdminCategoriesController.updateCategory);
 
-router.delete("/delete/:id", ensureAuthenticated, function (req, res) {
-  Category.remove(
-    {
-      _id: req.params.id
-    },
-    function (err, category) {
-      if (err) res.send(err);
-
-      console.log("Category deleted");
-      res.redirect("/categories?alert=deleted");
-    }
-  );
-});
-router.put("/update/:id", ensureAuthenticated, function (req, res) {
-  // use our category model to find the category we want
-  Category.findById(req.params.id, function (err, category) {
-    if (err) res.send(err);
-
-    category.name = req.body.name; // update the categories info
-
-    // save the category
-    category.save(function (err) {
-      if (err) res.send(err);
-
-      console.log("Category updated:", category);
-      res.redirect("/categories/edit/" + category._id + "?alert=deleted");
-    });
-  });
-});
 module.exports = router;
