+++ routes/admin/locations.js
@@ -1,101 +1,15 @@
-const Category = require("../../models/category");
-const Location = require("../../models/story");
-
 const express = require("express");
 const router = express.Router();
+const AdminLocationsController = require('../../controllers/admin/AdminLocationsController')
 
-router.get("/", async function(req, res) {
-  //here we get the whole collection and sort by order
-  let locations = await Location.find({})
-    .sort("order")
-    .exec();
-  let categories = await Category.find({}).exec();
+router.get("/", AdminLocationsController.getLocations);
 
-  res.render("admin/locations", {
-    locations: locations,
-    categories: categories,
-    message: res.locals.message,
-    color: res.locals.color
-  });
-});
+router.get("/:id", AdminLocationsController.getSingleLocation);
 
-router.get("/:id", function(req, res) {
-  Location.findById(req.params.id, function(err, story) {
-    res.render("story", {
-      story: story
-    });
-  });
-});
+router.get("/edit/:id", AdminLocationsController.editLocation);
 
-router.get("/edit/:id", function(req, res) {
-  Location.findById(req.params.id, async function(err, story) {
-    let allcategories = await Category.find({}).exec();
-    all = allcategories.map(cat => {
-      let match = story.categories
-        .map(pcat => pcat.toString())
-        .includes(cat._id.toString());
+router.post("/", AdminLocationsController.createLocation);
 
-      if (match) {
-        return Object.assign({ selected: true }, cat._doc);
-      } else {
-        return cat._doc;
-      }
-    });
-
-    res.render("editLocation", {
-      story: story,
-      categories: all,
-      message: res.locals.message,
-      color: res.locals.color
-    });
-  });
-});
-
-router.post("/", function(req, res) {
-  var story = new Location(); // create a new instance of the story model
-  story.name = req.body.name; // set the locations name (comes from the request)
-  story.content = req.body.content; // set the locations name (comes from the request)
-  story.order = req.body.order; // set the locations name (comes from the
-  story._categories = req.body.categories; // set the locations name (comes from the
-
-  // save the story and check for errors
-  story.save(function(err) {
-    if (err) res.send(err);
-    console.log("Location created:", story);
-    res.redirect("/admin/locations?alert=created");
-  });
-});
-
-router.delete("/delete/:id", function(req, res) {
-  Location.remove(
-    {
-      _id: req.params.id
-    },
-    function(err, story) {
-      if (err) res.send(err);
-
-      console.log("Location deleted");
-      res.redirect("/admin/locations?alert=deleted");
-    }
-  );
-});
-router.put("/update/:id", function(req, res) {
-  // use our story model to find the story we want
-  Location.findById(req.params.id, function(err, story) {
-    if (err) res.send(err);
-
-    story.name = req.body.name; // update the locations info
-    story.content = req.body.content; // update the locations info
-    story.order = req.body.order; // update the locations info
-    story.categories = req.body.categories;
-
-    // save the story
-    story.save(function(err) {
-      if (err) res.send(err);
-
-      console.log("Location updated:", story);
-      res.redirect("/admin/locations/edit/" + story._id + "?alert=updated");
-    });
-  });
-});
+router.delete("/delete/:id", AdminLocationsController.deleteLocation);
+router.put("/update/:id", AdminLocationsController.updateLocation);
 module.exports = router;
