+++ routes/users.js
@@ -1,102 +1,15 @@
 var express = require('express');
 var router = express.Router();
-var passport = require('passport');
-var LocalStrategy = require('passport-local').Strategy;
+var UserController = require('../controllers/UserController');
 
-var User = require('../models/user');
+router.get('/register', UserController.renderRegister);
 
-// Register
-router.get('/register', function (req, res) {
-  res.render('register');
-});
+router.get('/login', UserController.renderLogin);
 
-// Login
-router.get('/login', function (req, res) {
+router.post('/register', UserController.register);
 
-  res.render('login');
+router.post( '/login', UserController.login);
 
-});
-
-// Register User
-router.post('/register', function (req, res) {
-  var name = req.body.name;
-  var email = req.body.email;
-  var username = req.body.username;
-  var password = req.body.password;
-  var password2 = req.body.password2;
-
-  // Validation
-  req.checkBody('name', 'Name is required').notEmpty();
-  req.checkBody('email', 'Email is required').notEmpty();
-  req.checkBody('email', 'Email is not valid').isEmail();
-  req.checkBody('username', 'Username is required').notEmpty();
-  req.checkBody('password', 'Password is required').notEmpty();
-  req.checkBody('password2', 'Passwords do not match').equals(req.body.password);
-
-  var errors = req.validationErrors();
-  if (errors) {
-    res.render('register', {
-      errors: errors
-    });
-  } else {
-    var newUser = new User({
-      name: name,
-      email: email,
-      username: username,
-      password: password
-    });
-    User.createUser(newUser, function (err, user) {
-      if (err) throw err;
-      console.log(user);
-    });
-    res.redirect('/users/login');
-  }
-});
-
-passport.use(
-  new LocalStrategy((username, password, done) => {
-    console.log("GRRR", username)
-    User.getUserByUsername(username, (err, user) => {
-      if (err) throw err;
-      if (!user) {
-        return done(null, false, { message: 'Unknown User' });
-      }
-      User.comparePassword(password, user.password, (err, isMatch) => {
-        if (err) throw err;
-        if (isMatch) {
-          return done(null, user);
-        } else {
-          return done(null, false, { message: 'Invalid password' });
-        }
-      });
-    });
-  })
-);
-
-passport.serializeUser(function (user, done) {
-  done(null, user.id);
-});
-
-passport.deserializeUser(function (id, done) {
-  User.getUserById(id, function (err, user) {
-    done(err, user);
-  });
-});
-
-router.post(
-  '/login',
-  passport.authenticate('local', {
-    successRedirect: '/',
-    failureRedirect: '/users/login'
-  }),
-  (req, res) => {
-    res.redirect('/');
-  }
-);
-
-router.get('/logout', (req, res) => {
-  req.logout();
-  res.redirect('/users/login');
-});
+router.get('/logout', UserController.logout);
 
 module.exports = router;
