+++ views/admin/editCategory.pug
@@ -1,0 +1,18 @@
+extends ../layout.pug
+block content
+  +navigation
+  .container.mt-5.text-left
+          
+    .row
+      .col
+        h1 Edit category
+        form.mb-3.form-signin(method='POST', action='/admin/categories/update/'+category._id+'?_method=PUT')
+          
+          .mb-3
+            label.sr-only(for='inputName') Name
+            input#inputName.form-control(name='name', type='text', placeholder='Name', value=category.name required='', autofocus='')
+            
+          button.btn.btn-lg.btn-primary.btn-block(type='submit') Update
+
+        //-to deploy this app I needed a helper. In .env ive set "https://tmy.menkent.uberspace.de/ultrablog/" and pass this in index.js:14 as a template helper per res.locals.domain to use here absolute paths
+        a.btn.btn-outline-primary(href="/admin/categories")="Back"
