+++ views/admin/editStory.pug
@@ -1,4 +1,4 @@
-extends layout.pug
+extends ../layout.pug
 block content
   +navigation
   .container.mt-5.text-left
