+++ views/admin/locations.pug
@@ -4,20 +4,20 @@ block content
   .container.mt-5.text-left
     .row
       .col-md-8
-        h2 Category route
+        h2 Location route
         ul.list-group
-          each category in categories
+          each location in locations
             li.list-group-item.mb-3
-              h2.mr-3=category.name
-              form.d-flex.justify-content-between(method='POST', action='/admin/categories/delete/'+category._id +'?_method=DELETE')
+              h2.mr-3=location.name
+              form.d-flex.justify-content-between(method='POST', action='/admin/locations/delete/'+location._id +'?_method=DELETE')
                 if !public
                   .d-flex.align-items-center
-                    a.btn.btn-outline-primary.mr-2(href="categories/edit/"+category._id) Edit
+                    a.btn.btn-outline-primary.mr-2(href="locations/edit/"+location._id) Edit
                     button.btn.btn-outline-danger(type='submit') delete
 
       .col-md-4
-        form.form-signin(method='POST', action='/admin/categories?alert=create')
-          h2 Create a new Category
+        form.form-signin(method='POST', action='/admin/locations?alert=create')
+          h2 Create a new Location
           
           .mb-3
             label.sr-only(for='inputName') Name
